// src/components/SectionSearchResult/types.ts
export interface MappedPayload {
    name: string
    id: number
    image: string | null
    imageBack: string | null
    height?: number
    weight?: number
    types: string[]
    stats: Stat[]
    abilities: Ability[]
}
// ...
export interface Ability {
    ability: Ability2
    is_hidden: boolean
    slot: number
}

export interface Ability2 {
    name: string
    url: string
}
// ...
export interface Stat {
    base_stat: number
    effort: number
    stat: Stat2
}

export interface Stat2 {
    name: string
    url: string
}